+++ Zilon.Client/Assets/Zilon/Scripts/Common/PlayerPersonCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2f65dd0ad6ffff8408511d1f262348cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
