+++ Zilon.Client/Assets/Zilon/Scripts/Models/SectorScene/SectorVM.cs
@@ -31,6 +31,8 @@ using Zilon.Core.Tactics.Spatial;
 // ReSharper disable once UnusedMember.Global
 internal class SectorVM : MonoBehaviour
 {
+    public PlayerPersonCreator PersonCreator;
+
     private readonly List<MapNodeVM> _nodeViewModels;
     private readonly List<ActorViewModel> _actorViewModels;
     private readonly List<ContainerVm> _containerViewModels;
@@ -91,7 +93,7 @@ internal class SectorVM : MonoBehaviour
 
     [NotNull] [Inject] private readonly IScoreManager _scoreManager;
 
-    [Inject] private IHumanActorTaskSource _humanActorTaskSource;
+    [Inject] private readonly IHumanActorTaskSource _humanActorTaskSource;
 
     [Inject(Id = "monster")] private readonly IActorTaskSource _monsterActorTaskSource;
 
@@ -120,8 +122,6 @@ internal class SectorVM : MonoBehaviour
     [NotNull]
     [Inject(Id = "show-trader-modal-command")]
     private readonly ICommand _showTraderModalCommand;
-
-    private bool _canMove;
 
     public SectorVM()
     {
@@ -297,7 +297,6 @@ internal class SectorVM : MonoBehaviour
         if (!blocked)
         {
             _playerState.HoverViewModel = (IMapNodeViewModel)sender;
-            _canMove = _moveCommand.CanExecute();
         }
     }
 
@@ -513,63 +512,7 @@ internal class SectorVM : MonoBehaviour
     {
         if (_humanPlayer.MainPerson == null)
         {
-            var inventory = new Inventory();
-
-            var evolutionData = new EvolutionData(_schemeService);
-
-            var defaultActScheme = _schemeService.GetScheme<ITacticalActScheme>(personScheme.DefaultAct);
-
-            var person = new HumanPerson(personScheme, defaultActScheme, evolutionData, survivalRandomSource, inventory);
-
-            _humanPlayer.MainPerson = person;
-
-
-            var classRoll = UnityEngine.Random.Range(1, 6);
-            switch (classRoll)
-            {
-                case 1:
-                    AddEquipmentToActor(person.EquipmentCarrier, 2, "short-sword");
-                    AddEquipmentToActor(person.EquipmentCarrier, 1, "steel-armor");
-                    AddEquipmentToActor(person.EquipmentCarrier, 3, "wooden-shield");
-                    break;
-
-                case 2:
-                    AddEquipmentToActor(person.EquipmentCarrier, 2, "battle-axe");
-                    AddEquipmentToActor(person.EquipmentCarrier, 3, "battle-axe");
-                    AddEquipmentToActor(person.EquipmentCarrier, 0, "highlander-helmet");
-                    break;
-
-                case 3:
-                    AddEquipmentToActor(person.EquipmentCarrier, 2, "bow");
-                    AddEquipmentToActor(person.EquipmentCarrier, 1, "leather-jacket");
-                    AddEquipmentToActor(inventory, "short-sword");
-                    AddResourceToActor(inventory, "arrow", 10);
-                    break;
-
-                case 4:
-                    AddEquipmentToActor(person.EquipmentCarrier, 2, "fireball-staff");
-                    AddEquipmentToActor(person.EquipmentCarrier, 1, "scholar-robe");
-                    AddEquipmentToActor(person.EquipmentCarrier, 0, "wizard-hat");
-                    AddResourceToActor(inventory, "mana", 15);
-                    break;
-
-                case 5:
-                    AddEquipmentToActor(person.EquipmentCarrier, 2, "pistol");
-                    AddEquipmentToActor(person.EquipmentCarrier, 0, "elder-hat");
-                    AddResourceToActor(inventory, "bullet-45", 5);
-
-                    AddResourceToActor(inventory, "packed-food", 1);
-                    AddResourceToActor(inventory, "water-bottle", 1);
-                    AddResourceToActor(inventory, "med-kit", 1);
-
-                    AddResourceToActor(inventory, "mana", 5);
-                    AddResourceToActor(inventory, "arrow", 3);
-                    break;
-            }
-
-            AddResourceToActor(inventory, "packed-food", 1);
-            AddResourceToActor(inventory, "water-bottle", 1);
-            AddResourceToActor(inventory, "med-kit", 1);
+            _humanPlayer.MainPerson = PersonCreator.CreatePlayerPerson();
         }
 
         var actor = new Actor(_humanPlayer.MainPerson, player, startNode);
@@ -603,48 +546,6 @@ internal class SectorVM : MonoBehaviour
     {
         _container.InstantiateComponentOnNewGameObject<GameOverEffect>(nameof(GameOverEffect));
         _humanActorTaskSource.CurrentActor.Person.Survival.Dead -= HumanPersonSurvival_Dead;
-    }
-
-    private void AddEquipmentToActor(Inventory inventory, string equipmentSid)
-    {
-        try
-        {
-            var equipmentScheme = _schemeService.GetScheme<IPropScheme>(equipmentSid);
-            var equipment = _propFactory.CreateEquipment(equipmentScheme);
-            inventory.Add(equipment);
-        }
-        catch (KeyNotFoundException)
-        {
-            Debug.LogError($"Не найден объект {equipmentSid}");
-        }
-    }
-
-    private void AddEquipmentToActor(IEquipmentCarrier equipmentCarrier, int slotIndex, string equipmentSid)
-    {
-        try
-        {
-            var equipmentScheme = _schemeService.GetScheme<IPropScheme>(equipmentSid);
-            var equipment = _propFactory.CreateEquipment(equipmentScheme);
-            equipmentCarrier[slotIndex] = equipment;
-        }
-        catch (KeyNotFoundException)
-        {
-            Debug.LogError($"Не найден объект {equipmentSid}");
-        }
-    }
-
-    private void AddResourceToActor(Inventory inventory, string resourceSid, int count)
-    {
-        try
-        {
-            var resourceScheme = _schemeService.GetScheme<IPropScheme>(resourceSid);
-            var resource = _propFactory.CreateResource(resourceScheme, count);
-            inventory.Add(resource);
-        }
-        catch (KeyNotFoundException)
-        {
-            Debug.LogError($"Не найден объект {resourceSid}");
-        }
     }
 
     private void ActorOnUsedAct(object sender, UsedActEventArgs e)
